+++ features/meetup/create/components/CreateForm/index.tsx
@@ -6,7 +6,7 @@ import useToggle from "@/hooks/useToggle";
 import { useToast } from "@/providers/toast-provider";
 import { cn } from "@/utils/cn";
 import { Modal } from "@/components/ui/Modals";
-import { meetupQueryKeys, usePostMeetup, useUploadMeetupImage } from "../../../queries";
+import { usePostMeetup, useUploadMeetupImage } from "../../../queries";
 import { MeetupCreateRequest } from "../../../types";
 import { getKakaoPlace } from "../../../apis";
 import { modalSizeStyle } from "../../../styles";
@@ -21,6 +21,7 @@ import StepInfo from "../StepInfo";
 import StepSchedule from "../StepSchedule";
 import StepDesc from "../StepDesc";
 import { mypageQueryKeys } from "@/features/shared/queryKeys/mypage";
+import { meetupQueryKeys } from "@/features/shared/queryKeys/meetup";
 
 export type OnSubmit = (data: MeetupCreateRequest) => Promise<void>;
 export type OnSuccess = (id: number) => void;
